+++ amplitudelib/xs.cpp
@@ -20,25 +20,25 @@
 struct Inthelper_hadronprod
 {   
     AmplitudeLib *N;
-    REAL y, pt, xf;
-    REAL sqrts;
-    REAL miny,maxy;
-    REAL minpt,maxpt;
+    double y, pt, xf;
+    double sqrts;
+    double miny,maxy;
+    double minpt,maxpt;
     PDF* pdf;
     FragmentationFunction* frag;
     bool deuteron;
     Hadron final;
 };
 
-REAL Inthelperf_hadronprod(REAL z, void *p)
+double Inthelperf_hadronprod(double z, void *p)
 {
     // x1: fraction of parent hadron momentum carried by quark,
     // x2: Bjorken x for the dense system
     Inthelper_hadronprod* par = (Inthelper_hadronprod*)p;
-    REAL x1 = par->xf/z; REAL x2 = x1*std::exp(-2.0*par->y);
+    double x1 = par->xf/z; double x2 = x1*std::exp(-2.0*par->y);
     if (x1>1) return 0;
     
-    REAL y_A = std::log(par->N->X0()/x2);
+    double y_A = std::log(par->N->X0()/x2);
 
     if (y_A<0)
     {
@@ -50,13 +50,13 @@ REAL Inthelperf_hadronprod(REAL z, void *p)
     bool deuteron = par->deuteron;
     par->N->InitializeInterpolation(y_A);
 
-    REAL result = 0;
+    double result = 0;
 
     // Quark from proton:
     // UGD in fundamental representation
-    REAL nf = par->N->S_k(par->pt/z, y_A);
+    double nf = par->N->S_k(par->pt/z, y_A);
     // PDF and fragmentation
-    REAL xqf = par->pdf->xq(x1, par->pt, U)*par->frag->Evaluate(U, par->final, z, par->pt)
+    double xqf = par->pdf->xq(x1, par->pt, U)*par->frag->Evaluate(U, par->final, z, par->pt)
         + par->pdf->xq(x1, par->pt, D)*par->frag->Evaluate(D, par->final, z, par->pt);
 
     if (deuteron)
@@ -69,7 +69,7 @@ REAL Inthelperf_hadronprod(REAL z, void *p)
     result = nf*xqf;
 
     // Adjoint representation, gluon scatters
-    REAL na = par->N->S_k(par->pt/z, y_A, true);
+    double na = par->N->S_k(par->pt/z, y_A, true);
     xqf = par->pdf->xq(x1, par->pt, G)*par->frag->Evaluate(G, par->final, z, par->pt);
     if (deuteron) xqf *= 2.0;   // gluon pdf gets multiplied by 2
     result += na*xqf;
@@ -83,12 +83,12 @@ REAL Inthelperf_hadronprod(REAL z, void *p)
  * => use isospin symmetry
  * Default final state particle is PI0
  */
-REAL AmplitudeLib::dHadronMultiplicity_dyd2pt(REAL y, REAL pt, REAL sqrts,
+double AmplitudeLib::dHadronMultiplicity_dyd2pt(double y, double pt, double sqrts,
     FragmentationFunction* fragfun, PDF* pdf, bool deuteron, Hadron final )
 {
-    const REAL K = 1.0; // normalization factor
+    const double K = 1.0; // normalization factor
     // We assume light hadrons
-    REAL xf = pt/sqrts*std::exp(y);
+    double xf = pt/sqrts*std::exp(y);
 
     Inthelper_hadronprod helper;
     helper.N=this; helper.y=y; helper.pt=pt; helper.xf=xf;
@@ -96,7 +96,7 @@ REAL AmplitudeLib::dHadronMultiplicity_dyd2pt(REAL y, REAL pt, REAL sqrts,
     helper.final=final;
     helper.pdf=pdf; helper.frag=fragfun;
 
-    REAL result=0; REAL abserr=0;
+    double result=0; double abserr=0;
     const int MULTIPLICITYXINTPOINTS=10;
 
     
@@ -126,13 +126,13 @@ REAL AmplitudeLib::dHadronMultiplicity_dyd2pt(REAL y, REAL pt, REAL sqrts,
 /*
  * Multiplicity integrated over rapidity and p_T range
  */
-REAL Inthelperf_hadronprod_yint(REAL y, void *p);
-REAL Inthelperf_hadronprod_ptint(REAL pt, void *p);
+double Inthelperf_hadronprod_yint(double y, void *p);
+double Inthelperf_hadronprod_ptint(double pt, void *p);
 const int HADRONPROD_YINTPOINTS=3;
 const int HADRONPROD_PTINTPOINTS=3;
-const REAL HADRONPROD_INTACCURACY=0.01;
+const double HADRONPROD_INTACCURACY=0.01;
 
-REAL AmplitudeLib::HadronMultiplicity(REAL miny, REAL maxy, REAL minpt, REAL maxpt, REAL sqrts,
+double AmplitudeLib::HadronMultiplicity(double miny, double maxy, double minpt, double maxpt, double sqrts,
             FragmentationFunction *fragfun, PDF* pdf, bool deuteron, Hadron final )
 {
     Inthelper_hadronprod helper;
@@ -152,7 +152,7 @@ REAL AmplitudeLib::HadronMultiplicity(REAL miny, REAL maxy, REAL minpt, REAL max
     gsl_integration_workspace *workspace 
      = gsl_integration_workspace_alloc(HADRONPROD_YINTPOINTS);
 
-    int status=0; REAL abserr, result;
+    int status=0; double abserr, result;
     status=gsl_integration_qag(&fun, miny, maxy,
             0, HADRONPROD_INTACCURACY, HADRONPROD_YINTPOINTS,
             GSL_INTEG_GAUSS15, workspace, &result, &abserr);
@@ -168,7 +168,7 @@ REAL AmplitudeLib::HadronMultiplicity(REAL miny, REAL maxy, REAL minpt, REAL max
     return result;
 }
 
-REAL Inthelperf_hadronprod_yint(REAL y, void* p)
+double Inthelperf_hadronprod_yint(double y, void* p)
 {
     Inthelper_hadronprod* par = (Inthelper_hadronprod*)p;
     cout << "# yint y=" << y << endl;
@@ -178,7 +178,7 @@ REAL Inthelperf_hadronprod_yint(REAL y, void* p)
     par->y = y;
     fun.params=par;
     
-    int status=0; REAL abserr, result;
+    int status=0; double abserr, result;
     gsl_integration_workspace *workspace 
         = gsl_integration_workspace_alloc(HADRONPROD_PTINTPOINTS);
     status=gsl_integration_qag(&fun, par->minpt, par->maxpt,
@@ -197,7 +197,7 @@ REAL Inthelperf_hadronprod_yint(REAL y, void* p)
 
 }
 
-REAL Inthelperf_hadronprod_ptint(REAL pt, void* p)
+double Inthelperf_hadronprod_ptint(double pt, void* p)
 {
     Inthelper_hadronprod* par = (Inthelper_hadronprod*)p;
     return pt*par->N->dHadronMultiplicity_dyd2pt(par->y, pt, par->sqrts, par->frag,
@@ -212,29 +212,29 @@ REAL Inthelperf_hadronprod_ptint(REAL pt, void* p)
  */
 struct Inthelper_dps
 {
-    REAL miny, maxy;
-    REAL minpt1, minpt2;
-    REAL sqrts;
+    double miny, maxy;
+    double minpt1, minpt2;
+    double sqrts;
     FragmentationFunction* fragfun;
     bool deuteron;
     Hadron final;
     AmplitudeLib* N;
-    REAL y1,y2,pt1;
+    double y1,y2,pt1;
     PDF *pdf;
 };
-REAL Inthelperf_dps_y1(REAL y1, void *p);
-REAL Inthelperf_dps_y2(REAL y1, void *p);
-REAL Inthelperf_dps_pt1(REAL y1, void *p);
-REAL Inthelperf_dps_pt2(REAL y1, void *p);
+double Inthelperf_dps_y1(double y1, void *p);
+double Inthelperf_dps_y2(double y1, void *p);
+double Inthelperf_dps_pt1(double y1, void *p);
+double Inthelperf_dps_pt2(double y1, void *p);
 
 
 const int DPS_YINTPOINTS=1;
 const int DPS_PTINTPOINTS=2;
-const REAL DPS_YINTACCURACY=0.05;
-const REAL DPS_PTINTACCURACY=0.05;
-const REAL DPS_MAXPT=3.5;
+const double DPS_YINTACCURACY=0.05;
+const double DPS_PTINTACCURACY=0.05;
+const double DPS_MAXPT=3.5;
 
-REAL AmplitudeLib::DPS(REAL miny, REAL maxy, REAL minpt1, REAL minpt2, REAL sqrts,
+double AmplitudeLib::DPS(double miny, double maxy, double minpt1, double minpt2, double sqrts,
             FragmentationFunction* fragfun, bool deuteron, Hadron final)
 {
     Inthelper_dps helper;
@@ -253,7 +253,7 @@ REAL AmplitudeLib::DPS(REAL miny, REAL maxy, REAL minpt1, REAL minpt2, REAL sqrt
     gsl_integration_workspace *workspace 
      = gsl_integration_workspace_alloc(DPS_YINTPOINTS);
 
-    int status=0; REAL abserr, result;
+    int status=0; double abserr, result;
     /*status=gsl_integration_qag(&fun, miny, maxy,
             0, DPS_YINTACCURACY, DPS_YINTPOINTS,
             GSL_INTEG_GAUSS15, workspace, &result, &abserr);
@@ -271,7 +271,7 @@ REAL AmplitudeLib::DPS(REAL miny, REAL maxy, REAL minpt1, REAL minpt2, REAL sqrt
     
 }
 
-REAL Inthelperf_dps_y1(REAL y1, void* p)
+double Inthelperf_dps_y1(double y1, void* p)
 {
     Inthelper_dps* par = (Inthelper_dps*)p;
     par->y1=y1;
@@ -282,7 +282,7 @@ REAL Inthelperf_dps_y1(REAL y1, void* p)
     gsl_integration_workspace *workspace 
      = gsl_integration_workspace_alloc(DPS_YINTPOINTS);
 
-    int status=0; REAL abserr, result;
+    int status=0; double abserr, result;
     /*status=gsl_integration_qag(&fun, par->miny, par->maxy,
             0, DPS_YINTACCURACY, DPS_YINTPOINTS,
             GSL_INTEG_GAUSS15, workspace, &result, &abserr);
@@ -300,7 +300,7 @@ REAL Inthelperf_dps_y1(REAL y1, void* p)
     return result;
 }
 int ptint=0;
-REAL Inthelperf_dps_y2(REAL y2, void* p)
+double Inthelperf_dps_y2(double y2, void* p)
 {
     Inthelper_dps* par = (Inthelper_dps*)p;
     par->y2=y2;
@@ -311,7 +311,7 @@ REAL Inthelperf_dps_y2(REAL y2, void* p)
     gsl_integration_workspace *workspace 
      = gsl_integration_workspace_alloc(DPS_PTINTPOINTS);
 
-    int status; REAL abserr, result;
+    int status; double abserr, result;
     status=gsl_integration_qag(&fun, par->minpt1, DPS_MAXPT,
             0, DPS_PTINTACCURACY, DPS_PTINTPOINTS,
             GSL_INTEG_GAUSS15, workspace, &result, &abserr);
@@ -324,7 +324,7 @@ REAL Inthelperf_dps_y2(REAL y2, void* p)
     return result;
 }
 
-REAL Inthelperf_dps_pt1(REAL pt1, void* p)
+double Inthelperf_dps_pt1(double pt1, void* p)
 {
     Inthelper_dps* par = (Inthelper_dps*)p;
     par->pt1=pt1;
@@ -335,7 +335,7 @@ REAL Inthelperf_dps_pt1(REAL pt1, void* p)
     gsl_integration_workspace *workspace 
      = gsl_integration_workspace_alloc(DPS_PTINTPOINTS);
 
-    int status; REAL abserr, result;
+    int status; double abserr, result;
     status=gsl_integration_qag(&fun, par->minpt2, pt1,
             0, DPS_PTINTACCURACY, DPS_PTINTPOINTS,
             GSL_INTEG_GAUSS15, workspace, &result, &abserr);
@@ -349,11 +349,11 @@ REAL Inthelperf_dps_pt1(REAL pt1, void* p)
     return result*SQR(2.0*M_PI);    //(2\pi)^2: angular part
 }
 
-REAL Inthelperf_dps_pt2(REAL pt2, void* p)
+double Inthelperf_dps_pt2(double pt2, void* p)
 {
     Inthelper_dps* par = (Inthelper_dps*)p;
 
-    REAL n1,n2;
+    double n1,n2;
 
     //cout << "Evaluating pt1=" << par->pt1 << " pt2=" << pt2 << " y1=" << par->y1
     //<< " y2=" << par->y2 << endl;
